% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/P1-kang.R
\name{linear_rg}
\alias{linear_rg}
\title{Linear Equating Analytic Standard Error}
\usage{
linear_rg(x, y, ...)
}
\arguments{
\item{x, y}{frequency tables of ramdom group equating design}
}
\description{
This package is used to calcuate the analytic standard error
of linear equating (the delta method) under random groups design.
}
\details{
According to Kolen and Brennan (2014, p. 250), "the analytic method
(referred to as the delta method) that can be used to estimate the standard
errors of equating using sample statistics."
}
\examples{
#Data prep
library("equate")
data("ACTmath")

#Frequency tables of random group equating
x <- as.freqtab(ACTmath[, 1:2])
y <- as.freqtab(ACTmath[, c(1, 3)])

#Getting std. error of linear equating under random groups design
linear_rg(x,y)

}

